feat(event): add aggregation to events
Before only available in Documents and Ratings APIs,
aggregation is now added to Events API .

Same filters are applied to stats and list endpoints.
Moreover `groupBy` and `field` parameters can work with nested properties
such as event `metadata` or `object`,
which is important to aggregate all sort of data (core or custom events).