feat(email): timezone can be specified for date rendering
Date formatter from `intl-messageformat` is customized
to use the timezone if it's available.

The offset difference between the source timezone
and the target timezone is computed so the date is
modified before rendering.
The rendered date will be expressed in the local time
from the provided timezone, instead of server instance
local time.