feat: verify and decrypt an authentication information
Up to now, external systems cannot determine if a Stelace API key
or a user token is valid or not.

If some custom logic is implemented in external systems
(e.g. serverless architecture),
this can be helpful to restrict users who are able to
trigger this logic by inspecting the authentication
information
(e.g. some transaction logic can only be triggered by users
that are part of the transaction).