perf!: add TimescaleDB extension to PostgreSQL
Some log tables are intended to be very large due to their nature
(e.g. events, webhook/workflow logs).

Even with indices optimization and partitioning,
PostgreSQL can't beat TimescaleDB in terms of performance
(the latter is optimized for time-series):
https://blog.timescale.com/blog/timescaledb-vs-6a696248104e

The good point is that TimescaleDB is an extension of PostgreSQL
so it will benefit from existing tools
and the codebase migration won't be too massive.