chore: deprecate event.getStats
This endpoint is very flexible to get customized aggregations.
However the drawback is the performance if the table becomes large.
Due to this flexibility, it's very hard to optimize to get quick response.

In cases where results number is still very high after filters,
the response may be returned after few minutes.