refactor!: rework workflow types for clarity
Previously there was only three types of workflow log:
- notification
- action
- runError

Any information concerning skipping or stopping the workflow step
was embedded into workflow log step object.

New types 'skipped' and 'stopped' are introduced to avoid checking
into this nested object.

Moreover, multiple variables had to be checked to determine
the cause of a log error. Now we distinguish two types of errors:
- preRunError (e.g. incorrect expressions), renaming from old runError
- "new" runError, replacing action with step.error = true

If the property `notifyUrl` of a workflow object is specified,
the object sent to the URL will have some property changes:
- `status` removed
- `type` added