feat: add SSL options to PostgreSQL connection (#379)
Some database providers only accepts requests with SSL enabled
(e.g. Timescale Cloud).

We support the following SSL options: ssl, cert, key, ca
Those options will map from Knex.js to the library
https://github.com/iceddev/pg-connection-string
thanks to the Knex.js PR
https://github.com/knex/knex/pull/3410

Please be careful when configuring SSL connection.
Passing key value into cert or inversely can trigger
an uncaught exception and even `process.on('uncaughtException', fn)`
cannot catch it.

* docs: add instructions to configure SSL for production

* refactor: merge Docker PostgreSQL images to keep only SSL one

* test: use multi-domains certificate

for local and CircleCI environments.

Previously, test was failing with a single-domain certificate.
SSL certificate was rejected for CircleCI as the host
didn't match the certificate information (localhost).

For tests, we used a self-signed multi-domains certificate
to check database SSL connections.
However generating a self-signed multi-domain certificate isn't enough
for Node.js, as it will reject it because it is self-signed.
To overcome this, we have to sign it with
another root certificate (that can be self-signed).

For Knex connection configuration, we need to provide
the two certificates as `sslcert` and `sslca` properties.
In real world usage, you may only need to specify one of
them.